+++ export/yolo.py
@@ -10,7 +10,7 @@ including TensorRT for optimized GPU inference.
 
 Example:
     # Using local weights files:
-    python export_yolo.py \
+    python export/yolo.py \
         --det_weights yolov5n.pt \
         --imgsz 640 \
         --infsz 320 \
@@ -20,7 +20,7 @@ Example:
         --fname yolo.onnx
 
     # Using W&B artifacts:
-    python export_yolo.py \
+    python export/yolo.py \
         --det_weights YOLOv5n-IR:latest \
         --imgsz 640 \
         --batch-size 2 \
@@ -32,9 +32,16 @@ NOTE: For TensorRT 7 compatible models, use the --trt7-compatible flag.
 """
 
 import argparse
+import sys
+from pathlib import Path
 from typing import Tuple
 
-from export_utils import export_model_to_onnx, get_weights_path, transform_sz
+FILE = Path(__file__).resolve()
+ROOT = FILE.parents[1]  # YOLOv5 root directory
+if str(ROOT) not in sys.path:
+    sys.path.append(str(ROOT))  # add ROOT to PATH
+
+from utils.export import export_model_to_onnx, get_weights_path, transform_sz
 from models.custom import SeaYOLO
 
 def main(
